Move validation to mutation part
