+++ saleor/checkout/checkout_cleaner.py
@@ -1,14 +1,20 @@
-from typing import TYPE_CHECKING, Iterable, Optional, Type, Union
+from datetime import date
+from typing import TYPE_CHECKING, Iterable, List, Optional, Type, Union
 
+import graphene
 from django.core.exceptions import ValidationError
 
+from ..core.exceptions import GiftCardNotApplicable
+from ..core.taxes import TaxError
 from ..discount import DiscountInfo
+from ..giftcard.models import GiftCard
 from ..payment import gateway
 from ..payment import models as payment_models
 from ..payment.error_codes import PaymentErrorCode
 from ..plugins.manager import PluginsManager
 from . import models
 from .error_codes import CheckoutErrorCode, OrderFromCheckoutCreateErrorCode
+from .models import Checkout
 from .utils import clear_delivery_method, is_fully_paid, is_shipping_required
 
 if TYPE_CHECKING:
@@ -100,4 +106,84 @@ def validate_checkout_email(checkout: models.Checkout):
         raise ValidationError(
             "Checkout email must be set.",
             code=CheckoutErrorCode.EMAIL_NOT_SET.value,
+        )
+
+
+def _validate_gift_cards(checkout: Checkout):
+    """Check if all gift cards assigned to checkout are available."""
+    today = date.today()
+    all_gift_cards = GiftCard.objects.filter(checkouts=checkout.token).count()
+    active_gift_cards = (
+        GiftCard.objects.active(date=today).filter(checkouts=checkout.token).count()
+    )
+    if not all_gift_cards == active_gift_cards:
+        msg = "Gift card has expired. Order placement cancelled."
+        raise GiftCardNotApplicable(msg)
+
+
+def validate_checkout(
+    checkout_info: "CheckoutInfo",
+    lines: Iterable["CheckoutLineInfo"],
+    unavailable_variant_pks: Iterable[int],
+    discounts: List["DiscountInfo"],
+    manager: "PluginsManager",
+):
+    """Validate all required data for converting checkout into order."""
+    if not checkout_info.channel.is_active:
+        raise ValidationError(
+            {
+                "channel": ValidationError(
+                    "Cannot complete checkout with inactive channel.",
+                    code=OrderFromCheckoutCreateErrorCode.CHANNEL_INACTIVE.value,
+                )
+            }
+        )
+    if unavailable_variant_pks:
+        not_available_variants_ids = {
+            graphene.Node.to_global_id("ProductVariant", pk)
+            for pk in unavailable_variant_pks
+        }
+        code = OrderFromCheckoutCreateErrorCode.UNAVAILABLE_VARIANT_IN_CHANNEL.value
+        raise ValidationError(
+            {
+                "lines": ValidationError(
+                    "Some of the checkout lines variants are unavailable.",
+                    code=code,
+                    params={"variants": not_available_variants_ids},
+                )
+            }
+        )
+    if not lines:
+        raise ValidationError(
+            {
+                "lines": ValidationError(
+                    "Cannot complete checkout without lines",
+                    code=OrderFromCheckoutCreateErrorCode.NO_LINES.value,
+                )
+            }
+        )
+
+    if checkout_info.checkout.voucher_code and not checkout_info.voucher:
+        raise ValidationError(
+            {
+                "voucher_code": ValidationError(
+                    "Voucher not applicable",
+                    code=OrderFromCheckoutCreateErrorCode.VOUCHER_NOT_APPLICABLE.value,
+                )
+            }
+        )
+    validate_checkout_email(checkout_info.checkout)
+
+    clean_billing_address(checkout_info, OrderFromCheckoutCreateErrorCode)
+    clean_checkout_shipping(checkout_info, lines, OrderFromCheckoutCreateErrorCode)
+    _validate_gift_cards(checkout_info.checkout)
+
+    # call plugin's hooks to validate if we are able to create an order
+    # can raise TaxError
+    try:
+        manager.preprocess_order_creation(checkout_info, discounts, lines)
+    except TaxError as tax_error:
+        raise ValidationError(
+            "Unable to calculate taxes - %s" % str(tax_error),
+            code=OrderFromCheckoutCreateErrorCode.TAX_ERROR.value,
         )
